feat: new usePreviousValueEffect hook, profile page now showing articles
